MDEV-13727 rpl.rpl_concurrency_error failed
The test actually revealed a flaw in MDEV-8305
which inadvertently enrolled the trigger and
stored function into slow query reporting aimed exclusively
to the stored procedure.
Specifically to the test, a query on the master was logged
with a timestamp of the query's top-level statement but its (post
update) trigger computed one more (later) timestamp which got
inserted into another table.

Master-vs-slave discrepancy became evident thanks to
different execution time of the trigger combined with the fact of the
logged master timestamp in the micro-second fraction format was
truncated on the slave. The latter means the slave execution of the query
always started earlier than on the master as far as the timestamp was concerned.
And when the fractional part was close to 1 the 2nd timestamp acquired by
MDEV-8305 piece of logics could overflow to next integer value.
That's how the master table appeared to be temporarly behind the slave's one.

Fixed with restoring the pre-MDEV-8305 behavior not to reset
the current statement's timestamp when the query is not from a stored procedure.
That is the trigger's or stored function's queries continue to use
the top-level statement's timestamp as they are supposed to.